Add explicit cast of SInt32 to int to squelch warning.

+++ JFBCrypt.m
@@ -835,7 +835,7 @@ signed char index_64[] = {
 		[hashedPassword appendString: @"0"];
 	}
 	
-	[hashedPassword appendFormat: @"%d", rounds];
+	[hashedPassword appendFormat: @"%d", (int) rounds];
 	[hashedPassword appendString: @"$"];
 	[hashedPassword appendString: [JFBCrypt encodeData: saltData
 											  ofLength: [saltData length]]];
@@ -866,7 +866,7 @@ signed char index_64[] = {
 		[salt appendString: @"0"];
 	}
 	
-	[salt appendFormat: @"%d", numberOfRounds];
+	[salt appendFormat: @"%d", (int) numberOfRounds];
 	[salt appendString: @"$"];
 	[salt appendString: [JFBCrypt encodeData: randomData
 									ofLength: [randomData length]]];
